Changed to Doubles

+++ main.kt
@@ -1,8 +1,8 @@
 fun main(args: Array<String>) {
 
-    val voltage: Long = args[0].toLong();
-    val voltageDrop: Long = args[1].toLong();
-    val currentInMilliAmps: Long = args[2].toLong();
+    val voltage: Double = args[0].toDouble();
+    val voltageDrop: Double = args[1].toDouble();
+    val currentInMilliAmps: Double = args[2].toDouble();
 
     val amps: Double = currentInMilliAmps * .001;
     println("Ohms Required: " + ((voltage-voltageDrop)/amps));
